Collapse SelfSubsciption into a single class that must execute before the bus is started
